Create Behavioral Code
It is a classical example of a state machine consisting of 4 states : Idle, Count Up, Count Down, Overflow.  The input signals are 'clk' (clock input), 'act' (enable option which controls transition from  Idle state to other states, 'h_l' (choice of  Up counting or Down counting), 'rst' ( reset option). The output signals are 'cnt' (4 bit counter output), 'ovflw' (Overflow flag). When 'act', 'rst' inputs are high it goes to the counting operation ( when 'h_l' is high it works as a Up Counter and when 'h_l' is low it works as a down counter) .