Test Bench for the 4 bit Synchronous Counter
Here , all the input signals are initially set to '0' which keeps it in the Idle state. Then after setting the 'act' and 'rst' signals to high state, 'h_l' is set as high to count up to the highest limit as an Up Counter and then it is set as low to observe the Down Counter operation, after which Overflow state is observed.